Fix : 페이지네이션 컴포넌트 재렌더링 문제 해결 - AllItems 자식 컴포넌트에서 형제 컴포넌트로 수정
